backend/requirements.txt
# FastAPI Backend Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# NLP and Text Processing
spacy==3.7.2
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
numpy==1.24.3

# Document Processing
pdfplumber==0.10.3
python-docx==1.1.0

# Database
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.0

# Additional utilities
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
